Unified models classes to only ModelUtils
